Fix generales de bugs de JavaBeans (mayúsculas y minúsculas). Implementación de FeignClient entre micro servicios Verificación exitosa del flujo completo de integración de los micro servicios actuales en Postman.

+++ 4. anden/src/main/java/com/example/anden/Model/Anden.java
@@ -18,7 +18,7 @@ public class Anden {
 
     @Id
     @Column(nullable = false)
-    private int numeroanden;
+    private Long numeroanden;
 
     @Column(nullable = false)
     @NotBlank
